AQE, changes in pretained models

+++ image-retrieval-v1/jordi/main.py
@@ -97,14 +97,19 @@ def main(config):
             os.mkdir(model_dir)
 
 
-    ########### EXTRACT FEATURES IF NOT EXTRACTED PREVIUS #####################
+    ########### EXTRACT FEATURES IF NOT EXTRACTED PREVIOUS #####################
     #extract image features for any model only if not previously extracted
+    #also save features for AVERAGE QUERY EXPANSION if not calculated
     pending_models_extract = []
     for model_name in pretained_list_models:
         model_dir = os.path.join(work_dir, model_name)
         features_file = os.path.join(model_dir, 'features.pickle')
         if not os.path.isfile(features_file):
             pending_models_extract.append(model_name)
+        else:
+            features_file = os.path.join(model_dir, 'features_aqe.pickle')
+            if not os.path.isfile(features_file):
+                pending_models_extract.append(model_name)
 
     if len(pending_models_extract) > 0 :
         transform = transforms.Compose([
@@ -118,39 +123,60 @@ def main(config):
         train_loader = DataLoader(train_dataset, batch_size=config["batch_size"], shuffle=False)
 
         #create logfile for save statistics - extract features
-        fields = ['ModelName', 'DataSetSize','TransformsResize','ParametersCount', 'FeaturesSize', 'PostProcFeaturesSize', 'ProcessTime']
+        fields = ['ModelName', 'DataSetSize','TransformsResize','ParametersCount', 'FeaturesSize', 
+                    'PostProcFeaturesSize', 'ProcessTime','AverageQueryExpansionTopK']
         logfile = LogFile(fields)        
 
         #Create timer to calculate the process time
         proctimer = ProcessTime()
 
         for model_name in pending_models_extract:
+            proctimer.start()
             #PATH to save features
             model_dir = os.path.join(work_dir, model_name)
 
-            # Load pretrained model
-            pretained_model = pretained_models.load_pretrained_model(model_name)
-            if debug:print(pretained_model)
-            pretained_model.to(device)
-
-            proctimer.start()
-
-            #put the network into train mode and do a pass over the dataset without doing any backpropagation
-            pretained_model = pretained_models.tuning_batch_norm_statistics(pretained_model,train_loader) 
-            #extract features
-            features = pretained_models.extract_features(pretained_model,train_loader,transform) 
-            features_size = features[0].shape[0]
-            #save complete features
-            features_file = os.path.join(model_dir, 'features_full.pickle')
-            pickle.dump(features , open(features_file, 'wb'))
-
-            #normalize features            
-            features = pretained_models.postprocessing_features(features,config["PCAdimension"]) 
-            postproc_features_size = features[0].shape[0]
-
-            #save features
+            recalc_aqe = False 
+            #EXTRACT FEATURES IF FILE NOT EXISTS
             features_file = os.path.join(model_dir, 'features.pickle')
-            pickle.dump(features , open(features_file, 'wb'))
+            if not os.path.isfile(features_file):
+                recalc_aqe = True  #if generate new normalitzed features, allways generate AQE              
+                # Load pretrained model
+                pretained_model = pretained_models.load_pretrained_model(model_name)
+                if debug:print(pretained_model)
+                pretained_model.to(device)
+
+                #put the network into train mode and do a pass over the dataset without doing any backpropagation
+                pretained_model = pretained_models.tuning_batch_norm_statistics(pretained_model,train_loader) 
+                #extract features
+                features = pretained_models.extract_features(pretained_model,train_loader,transform) 
+                features_size = features[0].shape[0]
+                #save complete features
+                features_file = os.path.join(model_dir, 'features_full.pickle')
+                pickle.dump(features , open(features_file, 'wb'))
+
+                #normalize features            
+                features = pretained_models.postprocessing_features(features,config["PCAdimension"]) 
+                postproc_features_size = features[0].shape[0]
+
+                #save normalized features
+                features_file = os.path.join(model_dir, 'features.pickle')
+                pickle.dump(features , open(features_file, 'wb'))
+            else:
+                features = pickle.load(open(features_file , 'rb'))
+
+            #CALCULATE AVERAGE QUERY EXPANSION IF FILE NOT EXISTS
+            features_file = os.path.join(model_dir, 'features_aqe')
+            if not os.path.isfile(features_file) or recalc_aqe:
+                # average of the top-K ranked results (including the query itself)
+                num_queries = train_df.shape[0] #all dataset
+                #queries = create_ground_truth_queries( train_df, "FirstN", num_queries,[])
+                #q_indx, y_true = make_ground_truth_matrix(train_df, queries)
+                q_indx = range(0,num_queries - 1)
+                features_aqe = pretained_models.Average_Query_Expansion(features,q_indx,config["top_k_image"])
+                #save features - average query expansion
+                features_file = os.path.join(model_dir, 'features_aqe.pickle')
+                pickle.dump(features_aqe , open(features_file, 'wb'))
+
 
             #LOG
             processtime = proctimer.stop()
@@ -160,48 +186,113 @@ def main(config):
                         'ParametersCount':pretained_models.Count_Parameters(pretained_model), 
                         'FeaturesSize': features_size,
                         'PostProcFeaturesSize': postproc_features_size,
-                        'ProcessTime':processtime
+                        'ProcessTime':processtime,
+                        'AverageQueryExpansionTopK':config["top_k_image"]
                     } 
             logfile.writeLogFile(values)
+
 
         #Print and save logfile    
         logfile.printLogFile()
         logfile.saveLogFile_to_csv("extractfeatures",config)
+    ########### END EXTRACT FEATURES IF NOT EXTRACTED PREVIOUS #####################
 
 
-    ##### IMAGE RETRIEVAL TEST #############
+    ########### IMAGE RETRIEVAL TEST ###############################################
 
-    #create logfile for image retrieval
-    fields = ['ModelName', 'DataSetSize','FeaturesSize', 'ProcessTime', 'mAPqueries', 'mAP', 'Precision']
-    logfile = LogFile(fields)        
-    #Create timer to calculate the process time
-    proctimer = ProcessTime()
+    if  config["retrieval_test"] == "True":
+        #create logfile for image retrieval
+        fields = ['ModelName', 'DataSetSize', 'UsedFeatures', 'FeaturesSize', 'ProcessTime', 'PrecisionHits']
+        logfile = LogFile(fields)        
+        #Create timer to calculate the process time
+        proctimer = ProcessTime()
 
-    img_ds = MyDataset(dataset_image_dir,train_df)
+        img_ds = MyDataset(dataset_image_dir,train_df)
 
-    Test_size = 5  #number images to test
-    testid = np.random.randint(train_df.shape[0], size=Test_size)
-    it = 0
-    for index in testid:
-        # change this to the index of the image to use as a query!
-        imgidx = index
+        Test_size = config["retrieval_test_size"]  #number images to test
+        testid = np.random.randint(train_df.shape[0], size=Test_size)
+        it = 0
+        for index in testid:
+            # index of the image to use as a query!
+            imgidx = index
+            for model_name in pretained_list_models:
+                print(f'\rTest image ... {it}/{Test_size }', end='', flush=True)
+                proctimer.start()
+
+                model_dir = os.path.join(config["work_dir"], model_name)
+                features_file = os.path.join(model_dir , 'features.pickle')
+                features = pickle.load(open(features_file , 'rb'))
+
+                # Cosine Similarity
+                ranking = pretained_models.Cosine_Similarity(features,imgidx,config["top_k_image"])
+                Print_Similarity(img_ds,imgidx,ranking,model_name + " - COSINE SIMILARITY",debug)
+
+                # Cosine Similarity evaluation Hits
+                precision = evaluation_hits(train_df,imgidx,ranking)
+                print(f'\rPrecision hits COSINE SIMILARITY: {precision:0.04f}')
+
+                #LOG
+                processtime = proctimer.stop()
+                values = {  'ModelName':model_name, 
+                            'DataSetSize':train_df.shape[0],
+                            'UsedFeatures':'features.pickle',
+                            'FeaturesSize': features[0].shape[0],
+                            'ProcessTime': processtime,
+                            'PrecisionHits' : precision
+                        } 
+                logfile.writeLogFile(values)
+
+                proctimer.start()
+                # AVERAGE QUERY EXPANSION
+                features_file = os.path.join(model_dir , 'features_aqe.pickle')
+                features_aqe = pickle.load(open(features_file , 'rb'))
+
+                ranking = pretained_models.Cosine_Similarity(features_aqe,imgidx,config["top_k_image"])
+                Print_Similarity(img_ds,imgidx,ranking,model_name + " - AVERAGE QUERY EXPANSION",debug)
+
+                # Cosine Similarity evaluation Hits
+                precision = evaluation_hits(train_df,imgidx,ranking)
+                print(f'\rPrecision hits AVERAGE QUERY EXPANSION: {precision:0.04f}')
+                
+                #LOG
+                processtime = proctimer.stop()
+                values = {  'ModelName':model_name, 
+                            'DataSetSize':train_df.shape[0],
+                            'UsedFeatures':'features_aqe.pickle',
+                            'FeaturesSize': features_aqe[0].shape[0],
+                            'ProcessTime': processtime,
+                            'PrecisionHits' : precision
+                        } 
+                logfile.writeLogFile(values)
+            it += 1
+
+        #Print and save logfile    
+        logfile.printLogFile()
+        logfile.saveLogFile_to_csv("imageretrieval",config)
+
+    ########### END IMAGE RETRIEVAL TEST ###############################################
+
+
+    ########### EVALUATION ###############################################
+
+    if  config["evaluate"] == "True":
+        #create logfile for image retrieval
+        fields = ['ModelName', 'DataSetSize', 'UsedFeatures', 'FeaturesSize', 'ProcessTime', 'mAPqueries', 'mAP', 'PrecisionHits']
+        logfile = LogFile(fields)        
+        #Create timer to calculate the process time
+        proctimer = ProcessTime()
+
         for model_name in pretained_list_models:
-            print(f'\rTest image ... {it}/{Test_size }', end='', flush=True)
-            proctimer.start()
 
-            model_dir = os.path.join(config["work_dir"], model_name)
-            features_file = os.path.join(model_dir , 'features.pickle')
-            features = pickle.load(open(features_file , 'rb'))
+            for file in ['features.pickle','features_aqe.pickle']:
 
-            # COSINE SIMILARITY
-            ranking = pretained_models.Cosine_Similarity(features,imgidx,config["top_n_image"])
-            Print_Similarity(img_ds,imgidx,ranking,model_name + " - COSINE SIMILARITY",debug)
+                print(f'\rEvaluate File ... {file}', end='', flush=True)
+                proctimer.start()
 
-            #LOG
-            processtime = proctimer.stop()
-
-            if  config["evaluate"] == "True":
-                #Run evaluation 
+                # LOAD FEATURES
+                model_dir = os.path.join(config["work_dir"], model_name)
+                features_file = os.path.join(model_dir , file)
+                features = pickle.load(open(features_file , 'rb'))
 
                 #compute the similarity matrix
                 S = features @ features.T
@@ -213,29 +304,36 @@ def main(config):
 
                 #Compute mean Average Precision (mAP)
                 df = evaluate(S, y_true, q_indx)
-                if debug:print(df)
                 print(f'\rmAP: {df.ap.mean():0.04f}')
 
                 #Compute evaluation Hits
-                precision = evaluation_hits(train_df,imgidx,ranking)
-                print(f'\rPrecision hits: {precision:0.04f}')
+                accuracy = []
+                for index in q_indx:
+                    ranking = pretained_models.Cosine_Similarity(features,index,config["top_k_image"])
+                    precision = evaluation_hits(train_df,index,ranking)
+                    accuracy.append(precision)
+                precision = np.mean(accuracy)
+                print(f'\rPrecision Hits: {precision:0.04f}')
 
-                values = {  'ModelName':model_name, 
-                            'DataSetSize':train_df.shape[0], 
-                            'FeaturesSize': features[0].shape[0],
-                            'ProcessTime': processtime,
-                            'mAPqueries': num_queries,
-                            'mAP': f'mAP: {df.ap.mean():0.04f}',
-                            'Precision:' : precision
-                        } 
+                #LOG
+                processtime = proctimer.stop()
+                values = {'ModelName':model_name, 
+                        'DataSetSize':train_df.shape[0],
+                        'UsedFeatures': file,
+                        'FeaturesSize': features[0].shape[0],
+                        'ProcessTime': processtime,
+                        'mAPqueries': num_queries,
+                        'mAP': f'mAP: {df.ap.mean():0.04f}',
+                        'PrecisionHits:' : precision
+                    } 
                 logfile.writeLogFile(values)
+        #Print and save logfile    
+        logfile.printLogFile()
+        logfile.saveLogFile_to_csv("evaluation",config)
 
-            it += 1
 
-    #Print and save logfile    
-    logfile.printLogFile()
-    logfile.saveLogFile_to_csv("imageretrieval",config)
-    ##### END IMAGE RETRIEVAL TEST #############
+    ########### END EVALUATION ###############################################
+
 
 
     
@@ -245,15 +343,17 @@ if __name__ == "__main__":
         "dataset_labels_file" : "/home/manager/upcschool-ai/data/FashionProduct/styles.csv",
         "work_dir" : "/home/manager/upcschool-ai/data/FashionProduct/processed_datalab/",
         "transforms_resize" : 332,
-        "PCAdimension" : 512,
+        "PCAdimension" : 10,
         "train_size" : "all",  # "all" / "divide"=train(60%), Eval and test (20%) / number=fixed size
         "test_validate_size": 1, #used only for train_size = fixed zize
         "batch_size" : 8,
-        "top_n_image" : 10,  #multiple of 5
-        "mAP_n_queries": 1,
         "log_dir" : "/home/manager/upcschool-ai/data/FashionProduct/processed_datalab/log/",
-        "debug" : "False",
-        "evaluate" : "True" # True/False  Process Evaluation
+        "retrieval_test" : "True", # True/False  Retrieval Test
+        "retrieval_test_size" : 5,
+        "top_k_image" : 15,  #multiple of 5
+        "evaluate" : "False", # True/False  Process Evaluation
+        "mAP_n_queries": 300,
+        "debug" : "False"
     }
 
     main(config)
